% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_health.R
\name{check_cluster_health}
\alias{check_cluster_health}
\title{Check the cluster health}
\usage{
check_cluster_health(
  conn,
  index = NULL,
  expand_wildcards = "open",
  level = "cluster",
  local = FALSE,
  master_timeout = 30L,
  timeout = 30L,
  wait_for_active_shards = "0",
  wait_for_nodes = "<12",
  wait_for_events = "normal",
  wait_for_no_relocating_shards = FALSE,
  wait_for_no_initializing_shards = FALSE,
  wait_for_status = "green"
)
}
\arguments{
\item{conn}{`OpenSearchRConnection`, a connection object}

\item{index}{`character`, the name of the index}

\item{expand_wildcards}{`character`,}

\item{level}{`character`}

\item{local}{`logical`}

\item{master_timeout}{`integer`}

\item{timeout}{`integer`}

\item{wait_for_active_shards}{`character`}

\item{wait_for_nodes}{`character`}

\item{wait_for_events}{`character`}

\item{wait_for_no_relocating_shards}{`logical`}

\item{wait_for_no_initializing_shards}{`logical`}

\item{wait_for_status}{`character`}
}
\value{
`list`, the API response containing the status of the cluster
}
\description{
Use this function to check the health of the OpenSearch cluster.
}
