% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_int_to_str}
\alias{convert_int_to_str}
\title{Convert integer to character}
\usage{
convert_int_to_str(int, unit = "s")
}
\arguments{
\item{int}{`integer`, value to be converted}

\item{unit}{`character`, unit of time}
}
\value{
`character`, the integer as a converted string
}
\description{
This function converts an integer value to a string for the OpenSearch API.
Certain URL parameters like timeout values require units of time, i.e. "30s".
}
