% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_lgl_to_str}
\alias{convert_lgl_to_str}
\title{Convert logical to character}
\usage{
convert_lgl_to_str(lgl)
}
\arguments{
\item{lgl}{`logical`, value to be converted}
}
\value{
`character`, the logical as a converted string
}
\description{
This function converts an logical value to a string for the OpenSearch API.
Certain URL parameters like require boolean values to be specied as strings
i.e. "false" instead of `FALSE`.
}
