% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_connection.R
\name{create_connection}
\alias{create_connection}
\title{Create a connection object}
\usage{
create_connection(host, port, user, password)
}
\arguments{
\item{host}{`character`, the host url of the OpenSearch instance}

\item{port}{`integer`, the port number}

\item{user}{`character`, the username}

\item{password}{`character`, the password}
}
\value{
`OpenSearchRConnection`, a connection object
}
\description{
To interact with the OpenSearch REST API, the connection details are stored in
an object, which is then passed to functions that use the REST API.
}
