% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index.R
\name{create_index}
\alias{create_index}
\title{Create an index}
\usage{
create_index(conn, index)
}
\arguments{
\item{conn}{`OpenSearchRConnection`, a connection object}

\item{index}{`character`, the name of the index}
}
\value{
`logical`, TRUE if the index was created
}
\description{
The function creates an index. If the index already exists, the function
will throw an error.
}
