% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_index.R
\name{delete_index}
\alias{delete_index}
\title{Delete index}
\usage{
delete_index(conn, index)
}
\arguments{
\item{conn}{`OpenSearchRConnection`, a connection object}

\item{index}{`character`, the name of the index}
}
\value{
`list`, the content of the API response
}
\description{
This function deletes an index on OpenSearch. If the index does not exist
the function returns an error.
}
