% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints for OpenSearch}
\format{
An object of class \code{list} of length 1.
}
\usage{
endpoints
}
\description{
This object contains fixed endpoints to be used by the package to run
queries against.
}
\keyword{datasets}
