% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_exists.R
\name{index_exists}
\alias{index_exists}
\title{Check if index exists}
\usage{
index_exists(conn, index)
}
\arguments{
\item{conn}{`OpenSearchRConnection`, a connection object}

\item{index}{`character`, the name of the index}
}
\value{
`integer` the response status code
}
\description{
This function checks if the specified index exists, in which case the
status code of 200 is displayed.
}
