% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_status_code}
\alias{show_status_code}
\title{Get status code}
\usage{
show_status_code(response)
}
\arguments{
\item{response}{`response`, API response}
}
\value{
`integer`, status code
}
\description{
Returns the status code of an API response
}
