% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_exists.R
\name{index_exists}
\alias{index_exists}
\title{Check if index exists}
\usage{
index_exists(conn, index)
}
\arguments{
\item{conn}{`OpenSearchRConnection`, a connection object}

\item{index}{`character`, the name of the index}
}
\value{
`TRUE` if the index exists
}
\description{
This function checks if the specified index exists. If it does not exist,
an error is thrown.
}
