% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_response}
\alias{process_response}
\title{Get the content from a http response}
\usage{
process_response(response)
}
\arguments{
\item{response}{}
}
\description{
Get the content from a http response
}
